feat(suite-native): apply device authenticity settings in onboarding
