+++ suite-native/module-device-onboarding/src/screens/ConfirmFirmwareUpdateScreen.tsx
@@ -2,8 +2,7 @@ import { useSelector } from 'react-redux';
 
 import { useSetAtom } from 'jotai';
 
-import { SUPPORTS_DEVICE_AUTHENTICITY_CHECK } from '@suite-common/suite-constants';
-import { selectDeviceModel, selectHasDeviceFirmwareInstalled } from '@suite-common/wallet-core';
+import { selectHasDeviceFirmwareInstalled } from '@suite-common/wallet-core';
 import { selectIsDeviceFirmwareSupported } from '@suite-native/device';
 import {
     ConfirmFirmwareUpdateScreenContent,
@@ -17,7 +16,7 @@ import {
 
 import { updateOnboardingAnalyticsAtom } from '../../atoms';
 import { DeviceOnboardingScreenWithExitButton } from '../components/DeviceOnboardingScreenWithExitButton';
-
+import { useNavigateToNextScreenAfterFirmwareInstallation } from '../hooks/useNavigateToNextScreenAfterFirmwareInstallation';
 export const ConfirmFirmwareUpdateScreen = ({
     navigation,
 }: StackProps<
@@ -27,7 +26,9 @@ export const ConfirmFirmwareUpdateScreen = ({
     const hasDeviceFirmwareInstalled = useSelector(selectHasDeviceFirmwareInstalled);
     const isDeviceFirmwareSupported = useSelector(selectIsDeviceFirmwareSupported);
     const updateOnboardingAnalytics = useSetAtom(updateOnboardingAnalyticsAtom);
-    const deviceModel = useSelector(selectDeviceModel);
+
+    const { navigateToNextScreenAfterFirmwareInstallation } =
+        useNavigateToNextScreenAfterFirmwareInstallation();
 
     const handleUpdateConfirmation = () => {
         updateOnboardingAnalytics({
@@ -36,19 +37,11 @@ export const ConfirmFirmwareUpdateScreen = ({
         navigation.navigate(DeviceOnboardingStackRoutes.FirmwareInstallation);
     };
 
-    const supportsDeviceAuthentication = deviceModel
-        ? SUPPORTS_DEVICE_AUTHENTICITY_CHECK[deviceModel]
-        : true;
-
     const handleSkipUpdate = () => {
         updateOnboardingAnalytics({
             firmware: 'skip',
         });
-        navigation.navigate(
-            supportsDeviceAuthentication // TODO: check not only if supported, but also if allowed.
-                ? DeviceOnboardingStackRoutes.DeviceAuthenticity
-                : DeviceOnboardingStackRoutes.DeviceTutorial,
-        );
+        navigateToNextScreenAfterFirmwareInstallation();
     };
 
     return (
